Rename PersonCardHandle to TaskCardHandle in AddCommandTest
